Refactor server into service modules
