+++ mastodon_service.py
@@ -1,0 +1,85 @@
+import base64
+from io import BytesIO
+from typing import Dict, List, Optional
+
+from mastodon import Mastodon
+
+
+def validate_accounts(config: Dict) -> Dict[str, str]:
+    """Validate Mastodon account configuration and return a map of errors."""
+    errors: Dict[str, str] = {}
+    accounts = config.get("mastodon", {}).get("accounts")
+    if not accounts:
+        errors["_general"] = "mastodon.accounts section missing or empty"
+        return errors
+
+    for name, info in accounts.items():
+        account_errors = []
+        instance = info.get("instance_url")
+        token = info.get("access_token")
+
+        if not instance:
+            account_errors.append("missing instance_url")
+        elif "mastodon.example" in instance:
+            account_errors.append("instance_url looks like a placeholder")
+
+        if not token:
+            account_errors.append("missing access_token")
+        elif token == "YOUR_TOKEN":
+            account_errors.append("access_token looks like a placeholder")
+
+        if account_errors:
+            errors[name] = "; ".join(account_errors)
+    return errors
+
+
+def create_clients(config: Dict, account_errors: Dict[str, str]):
+    """Create Mastodon clients for all configured accounts."""
+    clients = {}
+    accounts = config.get("mastodon", {}).get("accounts", {})
+    for name, info in accounts.items():
+        if name in account_errors:
+            continue
+        try:
+            clients[name] = Mastodon(
+                access_token=info["access_token"],
+                api_base_url=info["instance_url"],
+            )
+        except Exception as exc:
+            # Log error but continue creating other clients
+            print(f"Failed to init Mastodon client for {name}: {exc}")
+    return clients
+
+
+def post_to_mastodon(
+    account: str,
+    text: str,
+    media: Optional[List[str]] = None,
+    *,
+    clients: Dict[str, Mastodon],
+    account_errors: Dict[str, str],
+):
+    """Post a toot with optional media."""
+    if account in account_errors:
+        return {"error": "Account misconfigured"}
+    client = clients.get(account)
+    if not client:
+        return {"error": "Account not configured"}
+
+    media_ids = None
+    if media:
+        media_ids = []
+        for item in media:
+            try:
+                data = base64.b64decode(item)
+                uploaded = client.media_post(BytesIO(data), mime_type="application/octet-stream")
+                media_ids.append(uploaded.get("id"))
+            except Exception as exc:
+                return {"error": f"Media upload failed: {exc}"}
+
+    try:
+        client.status_post(text, media_ids=media_ids)
+    except Exception as exc:
+        return {"error": str(exc)}
+
+    return {"posted": True}
