+++ twitter_service.py
@@ -1,0 +1,100 @@
+import base64
+from io import BytesIO
+from typing import Dict, List, Optional
+
+import tweepy
+
+
+def validate_accounts(config: Dict) -> Dict[str, str]:
+    """Validate Twitter account configuration and return a map of errors."""
+    errors: Dict[str, str] = {}
+    accounts = config.get("twitter", {}).get("accounts")
+    if not accounts:
+        errors["_general"] = "twitter.accounts section missing or empty"
+        return errors
+
+    placeholders = {
+        "consumer_key": "YOUR_CONSUMER_KEY",
+        "consumer_secret": "YOUR_CONSUMER_SECRET",
+        "access_token": "YOUR_ACCESS_TOKEN",
+        "access_token_secret": "YOUR_ACCESS_TOKEN_SECRET",
+        "bearer_token": "YOUR_BEARER_TOKEN",
+    }
+
+    for name, info in accounts.items():
+        account_errors = []
+        for key in placeholders:
+            val = info.get(key)
+            if not val:
+                account_errors.append(f"missing {key}")
+            elif val == placeholders[key]:
+                account_errors.append(f"{key} looks like a placeholder")
+
+        if account_errors:
+            errors[name] = "; ".join(account_errors)
+    return errors
+
+
+def create_clients(config: Dict, account_errors: Dict[str, str]):
+    """Create Tweepy clients for all configured accounts."""
+    clients = {}
+    accounts = config.get("twitter", {}).get("accounts", {})
+    for name, info in accounts.items():
+        if name in account_errors:
+            continue
+        try:
+            auth = tweepy.OAuth1UserHandler(
+                info["consumer_key"],
+                info["consumer_secret"],
+                info["access_token"],
+                info["access_token_secret"],
+            )
+            api = tweepy.API(auth)
+            client = tweepy.Client(
+                bearer_token=info["bearer_token"],
+                consumer_key=info["consumer_key"],
+                consumer_secret=info["consumer_secret"],
+                access_token=info["access_token"],
+                access_token_secret=info["access_token_secret"],
+            )
+            clients[name] = {"client": client, "api": api}
+        except Exception as exc:
+            print(f"Failed to init Twitter client for {name}: {exc}")
+    return clients
+
+
+def post_to_twitter(
+    account: str,
+    text: str,
+    media: Optional[List[str]] = None,
+    *,
+    clients: Dict[str, Dict[str, any]],
+    account_errors: Dict[str, str],
+):
+    """Send a tweet with optional media."""
+    if account in account_errors:
+        return {"error": "Account misconfigured"}
+    info = clients.get(account)
+    if not info:
+        return {"error": "Account not configured"}
+
+    client = info["client"]
+    api = info["api"]
+
+    media_ids = None
+    if media:
+        media_ids = []
+        for item in media:
+            try:
+                data = base64.b64decode(item)
+                uploaded = api.media_upload(filename="media", file=BytesIO(data))
+                media_ids.append(uploaded.media_id)
+            except Exception as exc:
+                return {"error": f"Media upload failed: {exc}"}
+
+    try:
+        client.create_tweet(text=text, media_ids=media_ids)
+    except Exception as exc:
+        return {"error": str(exc)}
+
+    return {"posted": True}
